+++ source/DSCResources/DSC_xGroupResource/DSC_xGroupResource.psm1
@@ -2508,6 +2508,11 @@ function Clear-GroupMember
 
     .PARAMETER MemberAsPrincipal
         The member to add to the group as a principal.
+    .NOTES
+         There is an issue reported at https://github.com/PowerShell/PSDscResources/issues/82
+         If local group already has members from trusted forests/domains, the Add method fails
+         The exact reason of the failure is unknown at the moment and as a workaround try-catch
+         block is used to fallback to ADSI WinNT provider.
 #>
 function Add-GroupMember
 {
@@ -2525,7 +2530,18 @@ function Add-GroupMember
         $MemberAsPrincipal
     )
 
-    $Group.Members.Add($MemberAsPrincipal)
+    try
+    {
+        $Group.Members.Add($MemberAsPrincipal)
+    }
+    catch
+    {
+        Write-Verbose -Message $script:localizedData.PrincipalCollectionAddMethodException
+        Write-Verbose -Message $_.ToString()
+        Write-Verbose -Message $script:localizedData.WinNTProviderFallback
+        [ADSI] $adsiGroup = ("WinNT://$env:COMPUTERNAME/$($Group.Name),group")
+        $adsiGroup.Add("WinNT://$($MemberAsPrincipal.Sid)")
+    }
 }
 
 <#